init credentials bug fix (took my 6 hours)

+++ backend/src/controllers/notes.ts
@@ -2,13 +2,16 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 //! Get ALL Notes
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   try {
     // throw Error("BOOM!");
     // throw createHttpError(401);
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authenticatedUserId);
+    const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -31,13 +34,17 @@ export const createNote: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
+  const authenticatedUserId = req.session.userId;
   try {
+    assertIsDefined(authenticatedUserId);
+
     //title i zorunlu tuttuğumuz icin olmadığı durumda bir hata ekledik
     if (!title) {
       throw createHttpError(400, "Note must have a title!");
     }
 
     const newNote = await NoteModel.create({
+      userId: authenticatedUserId, // burada userId'yi buraya koyarak user'ın id'sini saklayarak herkesin notu sadece kendi notlarını oluşturabilir
       title: title,
       text: text,
     });
@@ -51,7 +58,10 @@ export const createNote: RequestHandler<
 
 export const getSingleNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
   try {
+    assertIsDefined(authenticatedUserId);
+
     //! buraya normalde mongonun kabul ettiği 24 haneli idyi kontrol eden bir hata ekledik
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id!"); //400 bad request hatası veriyoruz
@@ -62,6 +72,10 @@ export const getSingleNote: RequestHandler = async (req, res, next) => {
     //! buraya da note olmadığı zamanki hatayı ekledik
     if (!note) {
       throw createHttpError(404, "Note not found!"); //404 not found hatası veriyoruz
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note!"); //401
     }
 
     res.status(200).json(note);
@@ -90,8 +104,11 @@ export const updateNote: RequestHandler<
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id!"); //400 bad request hatası veriyoruz
     }
@@ -102,6 +119,10 @@ export const updateNote: RequestHandler<
 
     if (!note) {
       throw createHttpError(404, "Note not found!"); //404 not found hatası veriyoruz
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note!"); //401
     }
 
     note.title = newTitle;
@@ -119,7 +140,11 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id!"); //400 bad request hatası veriyoruz
     }
@@ -127,6 +152,10 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, "Note not found!"); //404 not found hatası veriyoruz
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note!"); //401
     }
 
     const deletedNote = await NoteModel.findByIdAndDelete(noteId).exec();
